+++ source/standalone/demo/play_arms.py
@@ -19,7 +19,7 @@ From the default configuration file for these robots, zero actions imply a defau
 
 import argparse
 
-from omni.isaac.kit import SimulationApp
+from omni.isaac.orbit.app import AppLauncher
 
 # add argparse arguments
 parser = argparse.ArgumentParser("Welcome to Orbit: Omniverse Robotics Environments!")
@@ -28,9 +28,8 @@ parser.add_argument("--robot", type=str, default="franka_panda", help="Name of t
 args_cli = parser.parse_args()
 
 # launch omniverse app
-config = {"headless": args_cli.headless}
-simulation_app = SimulationApp(config)
-
+app_launcher = AppLauncher(headless=args_cli.headless)
+simulation_app = app_launcher.app
 
 """Rest everything follows."""
 
@@ -121,7 +120,7 @@ def main():
             break
         # If simulation is paused, then skip.
         if not sim.is_playing():
-            sim.step(render=not args_cli.headless)
+            sim.step(render=app_launcher.RENDER)
             continue
         # reset
         if ep_step_count % 1000 == 0:
@@ -145,7 +144,7 @@ def main():
         # apply action to the robot
         robot.apply_action(actions)
         # perform step
-        sim.step()
+        sim.step(render=app_launcher.RENDER)
         # update sim-time
         sim_time += sim_dt
         ep_step_count += 1