+++ source/standalone/demo/play_cloner.py
@@ -14,7 +14,7 @@ Reference: https://docs.omniverse.nvidia.com/app_isaacsim/app_isaacsim/tutorial_
 
 import argparse
 
-from omni.isaac.kit import SimulationApp
+from omni.isaac.orbit.app import AppLauncher
 
 # add argparse arguments
 parser = argparse.ArgumentParser("Welcome to Orbit: Omniverse Robotics Environments!")
@@ -24,9 +24,8 @@ parser.add_argument("--num_robots", type=int, default=128, help="Number of robot
 args_cli = parser.parse_args()
 
 # launch omniverse app
-config = {"headless": args_cli.headless}
-simulation_app = SimulationApp(config)
-
+app_launcher = AppLauncher(headless=args_cli.headless)
+simulation_app = app_launcher.app
 
 """Rest everything follows."""
 
@@ -146,7 +145,7 @@ def main():
             break
         # If simulation is paused, then skip.
         if not sim.is_playing():
-            sim.step(render=not args_cli.headless)
+            sim.step(render=app_launcher.RENDER)
             continue
         # reset
         if ep_step_count % 100 == 0:
@@ -169,7 +168,7 @@ def main():
         # apply actions
         robot.apply_action(actions)
         # perform step
-        sim.step()
+        sim.step(render=app_launcher.RENDER)
         # update sim-time
         sim_time += sim_dt
         ep_step_count += 1