+++ source/standalone/demo/play_quadrupeds.py
@@ -20,7 +20,7 @@ From the default configuration file for these robots, zero actions imply a stand
 
 import argparse
 
-from omni.isaac.kit import SimulationApp
+from omni.isaac.orbit.app import AppLauncher
 
 # add argparse arguments
 parser = argparse.ArgumentParser("Welcome to Orbit: Omniverse Robotics Environments!")
@@ -28,9 +28,8 @@ parser.add_argument("--headless", action="store_true", default=False, help="Forc
 args_cli = parser.parse_args()
 
 # launch omniverse app
-config = {"headless": args_cli.headless}
-simulation_app = SimulationApp(config)
-
+app_launcher = AppLauncher(headless=args_cli.headless)
+simulation_app = app_launcher.app
 
 """Rest everything follows."""
 
@@ -150,7 +149,7 @@ def main():
             break
         # If simulation is paused, then skip.
         if not sim.is_playing():
-            sim.step(render=not args_cli.headless)
+            sim.step(render=app_launcher.RENDER)
             continue
         # reset
         if count % 1000 == 0:
@@ -170,7 +169,7 @@ def main():
         robot_c.apply_action(actions)
         robot_a.apply_action(actions)
         # perform step
-        sim.step()
+        sim.step(render=app_launcher.RENDER)
         # update sim-time
         sim_time += sim_dt
         count += 1