+++ source/standalone/demo/play_ycb_objects.py
@@ -12,7 +12,7 @@ This script demonstrates how to use the rigid objects class.
 
 import argparse
 
-from omni.isaac.kit import SimulationApp
+from omni.isaac.orbit.app import AppLauncher
 
 # add argparse arguments
 parser = argparse.ArgumentParser("Welcome to Orbit: Omniverse Robotics Environments!")
@@ -20,9 +20,8 @@ parser.add_argument("--headless", action="store_true", default=False, help="Forc
 args_cli = parser.parse_args()
 
 # launch omniverse app
-config = {"headless": args_cli.headless}
-simulation_app = SimulationApp(config)
-
+app_launcher = AppLauncher(headless=args_cli.headless)
+simulation_app = app_launcher.app
 
 """Rest everything follows."""
 
@@ -130,7 +129,7 @@ def main():
             break
         # If simulation is paused, then skip.
         if not sim.is_playing():
-            sim.step(render=not args_cli.headless)
+            sim.step(render=app_launcher.RENDER)
             continue
         # reset
         if count % 250 == 0:
@@ -153,7 +152,7 @@ def main():
             rigid_object.reset_buffers()
             print(">>>>>>>> Reset!")
         # perform step
-        sim.step()
+        sim.step(render=app_launcher.RENDER)
         # update sim-time
         sim_time += sim_dt
         count += 1